Fix frame range rounding error on xmot import
The calculations done by the xmot import to determine the frame range of
the imported animation did not take into account the inherent inaccuracy
of floating point. This could result in the start and/or end of frame
range being off by one.

However, such an incorrect frame range has no effect on the animations
that are later exported, i.e., the generated .xmot files.